pyproject.toml
[project]
name = "tsujimoto"
version = "0.1.0"
description = "Neural networks for Japanese kanji character recognition using ETL9G dataset"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "paazmaya", email = "paazmaya@yahoo.com" }
]
keywords = ["kanji", "recognition", "machine-learning", "etl9g", "japanese", "character-recognition", "pytorch", "onnx"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.11"

dependencies = [
    # Core ML framework with CUDA support
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    # Data processing and utilities
    "numpy>=2.3.3",
    "scikit-learn>=1.7.2",
    "matplotlib>=3.10.7",
    "tqdm>=4.67.1",
    "opencv-python>=4.11.0",
    "Pillow>=11.3.0",
    # Model export and deployment (GPU-accelerated)
    "onnx>=1.19.1",
    "onnxruntime-gpu>=1.23.2",
    "onnxruntime-tools>=1.7.0",
    # CO2 emissions tracking
    "codecarbon>=3.0.7",
    # System monitoring
    "psutil>=7.1.0",
    # Data handling
    "safetensors>=0.7.0",
    "bitsandbytes>=0.48.2",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.1.1",
    "ruff>=0.14.5",
    "pytest>=8.4.2",
    "pytest-cov>=5.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.5",
    "pytest>=8.4.2",
]

[project.urls]
Homepage = "https://github.com/paazmaya/tsujimoto"
Repository = "https://github.com/paazmaya/tsujimoto"
Issues = "https://github.com/paazmaya/tsujimoto/issues"

# PyTorch with CUDA support requires special index
[[tool.uv.index]]
name = "torch"
url = "https://download.pytorch.org/whl/cu130"
explicit = true
packages = ["torch", "torchvision"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "ISC", "PIE", "T20", "Q"]
ignore = ["E501", "S101", "S311"]

# ML/data science code uses uppercase variable names for matrices and tensors (X, Y, K, Q, V, etc.)
[tool.ruff.lint.per-file-ignores]
"scripts/train_cnn_model.py" = ["N806"]
"scripts/train_vit.py" = ["N806"]
"scripts/train_qat.py" = ["N806"]
"scripts/train_hiercode.py" = ["N806"]
"scripts/train_radical_rnn.py" = ["N806"]
"scripts/train_hiercode_higita.py" = ["N806"]
"scripts/optimization_config.py" = ["N806"]
"scripts/load_multi_etl.py" = ["N806"]
"scripts/prepare_dataset.py" = ["N806"]
"scripts/convert_to_safetensors.py" = ["N806"]
"scripts/convert_to_onnx.py" = ["N806"]
"scripts/quantize_model.py" = ["N806"]
"scripts/test_etl9g_setup.py" = ["N806"]
"scripts/rnn/rnn_model.py" = ["N806"]
